requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
isodate==0.6.1
kombu==5.3.0
Markdown==3.4.3
mysqlclient==2.1.1
prompt-toolkit==3.0.38
pyparsing==3.0.9
python-decouple==3.8
pytz==2023.3
rdflib==6.3.2
six==1.16.0
SPARQLWrapper==2.0.0
sqlparse==0.4.4
typing_extensions==4.6.2
vine==5.0.0
wcwidth==0.2.6
